rename assetAnnotation
